benchmarks/kn/original/video-analytics/video-decoder/requirements.txt
parliament-functions==0.1.0
environs
pycapnp
opencv-python-headless
requests
